Add files via upload
this function takes in a list of three arrays with three values each. Each list represents a row of the tic tac toe board and the values in each list the columns. 0 for no move 1 for x and 2 for o and then returns values bases on the input. if the game hasn't finished it returns -1 if it's a tie 0 and if either x or o won returns the representing number